corrección de main.css a main.js

+++ main.js
@@ -4,7 +4,7 @@ numero1 = prompt("Por favor, ingresa un número");
 numero2 = prompt("Por favor, ingresa otro número");
 numero3 = prompt("Por favor, ingresa un último número");
 
-resultadoSuma = parseInt(numero1) + parseInt(numero2) + parseInt(numero3);
+resultadoSuma = parseFloat(numero1) + parseFloat(numero2) + parseFloat(numero3);
 promedio = resultadoSuma/3;
 
 alert(promedio);
